Speed up included hook, remove global list of receivers
This speeds up a plugin's included hook by changing the how it
detects if its been included or not. As part of this, the list of
receivers that each plugin stored has been removed as it is no
longer needed. Overall this doesn't change the core functionality
of much-plugin.

This changes how a plugin detects if its been included or not.
Previously, it would see if its been added to a list of receivers
that it stored and if not, it would add itself and run its hooks.
Now, it uses an included detector mixin that is unique to the
plugin. Since the plugin controls when this mixin is included, the
plugin can check in its included hook if its detector mixin has
been included. If not, include it and class eval its plugin hooks,
otherwise do nothing.

This makes plugin inclusion much faster and avoids a global list
of receivers that can build up. Typically, the list of receivers
should only be a handful of objects but if much-plugin is used with
an algorithm that creates anonymous classes then the list of
receivers can grow to a very large array. Furthermore, the array
keeps the classes from being garbage collected because it is always
holding a reference to them.

This also adds a bench script for comparing much-plugin's
performance with a regular mixin. Using much-plugin should always
be slightly slower but it should be a fairly insignificant
amount (about 1/10th of a millisecond on 10,000 iterations).